\name{LSspec-package}
\alias{LSspec-package}
\alias{LSspec}
\docType{package}
\title{
\packageTitle{LSspec}
}
\description{
\packageDescription{LSspec}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LSspec}
\packageIndices{LSspec}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{LSspec}

Maintainer: \packageMaintainer{LSspec}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
